+++ frontend/src/routes/map/utils/layers/index.ts
@@ -14,8 +14,7 @@ import type {
 	LineLayerSpecification,
 	SymbolLayerSpecification,
 	CircleLayerSpecification,
-	FilterSpecification,
-	ExpressionSpecification
+	FilterSpecification
 } from 'maplibre-gl';
 
 import { streetViewLineLayer, streetViewCircleLayer } from '$routes/map/utils/layers/street_view';
@@ -64,6 +63,7 @@ import {
 	createLinePatternLayer
 } from '$routes/map/utils/layers/vector/line_string';
 import { createCircleLayer } from '$routes/map/utils/layers/vector/point';
+import { getTemporalFilter } from '$routes/map/utils/layers/vector/filter';
 
 import { get } from 'svelte/store';
 
@@ -120,51 +120,6 @@ export interface LayerItem {
 	'source-layer'?: string;
 	filter?: FilterSpecification;
 }
-
-const getTemporalFilter = (entry: GeoDataEntry): FilterSpecification | undefined => {
-	if (entry.type !== 'vector') return undefined;
-	if (entry.format.type !== 'geojson') return undefined;
-
-	const temporalFilterState = entry.state?.temporalFilter;
-	if (!temporalFilterState?.enabled) return undefined;
-
-	const temporalConfig =
-		entry.properties.temporal ??
-		(entry.properties.attributeView.timeKey
-			? { key: entry.properties.attributeView.timeKey }
-			: undefined);
-	if (!temporalConfig) return undefined;
-
-	const temporalKeys = [temporalConfig.key, ...(temporalConfig.alternateKeys ?? [])].filter(
-		(key): key is string => Boolean(key)
-	);
-	if (temporalKeys.length === 0) return undefined;
-
-	const temporalValues = (temporalConfig.items ?? []).map((item) => item.raw);
-
-	if (temporalValues.length === 0) return undefined;
-
-	const startIndex = Math.min(temporalFilterState.startIndex, temporalValues.length - 1);
-	const endIndex = Math.min(temporalFilterState.endIndex, temporalValues.length - 1);
-	const startValue = temporalValues[startIndex];
-	const endValue = temporalValues[endIndex];
-	if (!startValue || !endValue) return undefined;
-
-	const temporalExpression =
-		temporalKeys.length === 1
-			? (['get', temporalKeys[0]] as ExpressionSpecification)
-			: ([
-					'coalesce',
-					...temporalKeys.map((key) => ['get', key]),
-					''
-				] as unknown as ExpressionSpecification);
-
-	return [
-		'all',
-		['>=', temporalExpression, startValue],
-		['<=', temporalExpression, endValue]
-	] as unknown as FilterSpecification;
-};
 
 export const createVectorLayer = (
 	layer: LayerItem,
